Adicionando USER e ADMIN como SCOPE, atualizando os controlleres e adicionando exceção

+++ src/main/java/bank/customer/controller/CustomerController.java
@@ -5,6 +5,7 @@ import bank.customer.service.CustomerServiceImpl;
 import bank.model.Customer;
 import bank.customer.request.CustomerRequest;
 
+import bank.security.exceptions.CustomAuthorizationException;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import jakarta.transaction.Transactional;
@@ -15,6 +16,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -33,30 +36,60 @@ public class CustomerController {
     }
     private static final Logger logger = Logger.getLogger(Customer.class.getName());
 
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
     /*Para todos os CLientes*/
     @ApiOperation(value ="Bring all Customer")
     @GetMapping
-    @PreAuthorize("hasAuthority('SCOPE_view_all_customers')")
+    @PreAuthorize("hasAuthority('SCOPE_admin')")
     public List<Customer> getAllCustomer() {
+        // Verificar se o usuário tem o escopo
+
+        boolean hasAdminScope = authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("SCOPE_view_customer"));
+
+        if (!hasAdminScope) {
+            throw new CustomAuthorizationException("Acesso negado");
+        }
+
         logger.info("Returning all costumer");
         return customerService.getAllCustomers();
     }
 
     @ApiOperation(value = "Bring a specific customer")
     @GetMapping("/{cpf}")
-    @PreAuthorize("hasAuthority('SCOPE_view_customer')")
-
+    @PreAuthorize("hasAuthority('SCOPE_admin')")
     public Optional<Customer> getCustomer(@PathVariable @Valid String cpf) {
+        // Verificar se o usuário tem o escopo
+        boolean hasAdminScope = authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("SCOPE_admin"));
+
+        if (!hasAdminScope) {
+            throw new CustomAuthorizationException("Acesso negado");
+        }
+
         logger.info("Returning a specific costumer");
         return Optional.ofNullable(customerService.getCustomerCpf(cpf));
     }
 
+
     /* Registro do Customer */
     @ApiOperation(value = "Customer Register")
     @PostMapping
-    @PreAuthorize("hasAuthority('SCOPE_register_customer')")
-
+    @PreAuthorize("hasAuthority('SCOPE_admin') or hasAuthority('SCOPE_user')")
     public ResponseEntity<Customer> registerCustomer (@RequestBody @Valid CustomerRequest customerRequest) {
+        // Verificar se o usuário tem o escopo
+        authentication.getAuthorities().stream()
+                .anyMatch(authority -> {
+                    authority.getAuthority();
+                    return false;
+                });
+        boolean hasAdminScope = false;
+
+        if (!hasAdminScope) {
+            throw new CustomAuthorizationException("Acesso negado");
+        }
+
         Customer customer = customerService.registerCustomer(customerRequest);
         logger.info("Costumer registered");
         return new ResponseEntity<>(customer, HttpStatus.CREATED);
@@ -65,9 +98,20 @@ public class CustomerController {
     /* Atualizar o Customer */
     @ApiOperation(value = "Customer Update")
     @PutMapping("/update")
-    @PreAuthorize("hasAuthority('SCOPE_update_customer')")
-
+    @PreAuthorize("hasAuthority('SCOPE_admin') or hasAuthority('SCOPE_user')")
     public ResponseEntity<Customer> updateCustomer (@RequestBody @Valid CustomerRequest cpf) {
+        // Verificar se o usuário tem o escopo
+        authentication.getAuthorities().stream()
+                .anyMatch(authority -> {
+                    authority.getAuthority();
+                    return false;
+                });
+        boolean hasAdminScope = false;
+
+        if (!hasAdminScope) {
+            throw new CustomAuthorizationException("Acesso negado");
+        }
+
         Customer customer = customerService.updateCustomer(cpf);
         logger.info("Costumer updated");
         return new ResponseEntity<Customer>(customer, HttpStatus.ACCEPTED);
@@ -77,8 +121,16 @@ public class CustomerController {
     @ApiOperation(value = "Deleting Customer")
     @Transactional
     @DeleteMapping("/{cpf}")
-    @PreAuthorize("hasAuthority('SCOPE_delete_customer')")
+    @PreAuthorize("hasAuthority('SCOPE_admin')")
     public ResponseEntity<?> deleteCustomer(@PathVariable String cpf) {
+        // Verificar se o usuário tem o escopo view_customer
+        boolean hasAdminScope = authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("SCOPE_admin"));
+
+        if (!hasAdminScope) {
+            throw new CustomAuthorizationException("Acesso negado");
+        }
+
         logger.info("Costumer deleted");
         customerService.deleteCustomer(cpf);
